+++ demo/settings.py
@@ -11,6 +11,7 @@ https://docs.djangoproject.com/en/4.0/ref/settings/
 """
 import os
 from pathlib import Path
+from importlib.util import find_spec
 from djplus.auth.settings import *
 
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
@@ -110,3 +111,11 @@ DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
 AUTH_PASSWORD_HASHERS = (
     "tests.auth.test_hashers.pbkdf2_hasher",
 )
+
+if find_spec("debug_toolbar"):
+    INSTALLED_APPS += ["debug_toolbar"]
+    MIDDLEWARE += ["debug_toolbar.middleware.DebugToolbarMiddleware"]
+    DEBUG_TOOLBAR_CONFIG = {
+        "SHOW_TEMPLATE_CONTEXT": True,
+    }
+    INTERNAL_IPS = ["127.0.0.1", "10.0.2.2"]