+++ demo/urls.py
@@ -15,8 +15,12 @@ Including another URLconf
 """
 # from django.contrib import admin
 from django.urls import path, include
+from django.conf import settings
 
 urlpatterns = [
     # path('admin/', admin.site.urls),
     path("auth/", include("djplus.auth.urls", namespace="auth")),
 ]
+
+if "debug_toolbar" in settings.INSTALLED_APPS:
+    urlpatterns += [path("__debug__/", include("debug_toolbar.urls", namespace="djdt"))]