+++ djplus/auth/middleware.py
@@ -8,13 +8,32 @@ class AuthenticationMiddleware:
         self.get_response = get_response
 
     def __call__(self, request: HttpRequest):
-        session_key = request.COOKIES.get(settings.AUTH_SESSION_COOKIE_NAME, None)
         try:
-            session = Session.objects.get(pk=session_key)
+            session = Session.objects.get(pk=request.COOKIES.get(settings.AUTH_SESSION_COOKIE_NAME, None))
         except Session.DoesNotExist:
-            request.session = None
+            request.session = Session()
             request.user = AnonymousUser()
         else:
             request.session = session
-            request.user = session.user
-        return self.get_response(request)
+            request.user = session.user or AnonymousUser()
+
+        response = self.get_response(request)
+
+        if request.session.is_empty:
+            request.session.delete()
+        else:
+            request.session.save()
+
+        if request.session.user:
+            response.set_cookie(
+                settings.AUTH_SESSION_COOKIE_NAME,
+                request.session.id,
+                max_age=settings.AUTH_SESSION_COOKIE_AGE,
+                expires=None,
+                domain=settings.AUTH_SESSION_COOKIE_DOMAIN,
+                path=settings.AUTH_SESSION_COOKIE_PATH,
+                secure=settings.AUTH_SESSION_COOKIE_SECURE,
+                httponly=settings.AUTH_SESSION_COOKIE_HTTPONLY,
+                samesite=settings.AUTH_SESSION_COOKIE_SAMESITE,
+            )
+        return response