+++ djplus/auth/urls.py
@@ -1,8 +1,10 @@
 from django.urls import path
-from .views import login
+from .views import login, logout, signup
 
 app_name = "auth"
 
 urlpatterns = [
     path("login/", login, name="login"),
+    path("logout/", logout, name="logout"),
+    path("signup/", signup, name="signup"),
 ]