+++ tests/auth/test_hashers.py
@@ -21,10 +21,8 @@ except ImportError:
 
 pbkdf2_hasher = PBKDF2PasswordHasher(iterations=1)
 scrypt_hasher = ScryptPasswordHasher(work_factor=2)
-if argon2:
-    argon2_hasher = Argon2PasswordHasher(memory_cost=8, time_cost=1, parallelism=1)
-if bcrypt:
-    bcrypt_hasher = BcryptPasswordHasher(rounds=4)
+argon2_hasher = Argon2PasswordHasher(memory_cost=8, time_cost=1, parallelism=1)
+bcrypt_hasher = BcryptPasswordHasher(rounds=4)
 
 
 class BasePasswordHasherTest(TestCase):