+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0] - 2022-07-25
+### Added
+- add `argon2-cffi` and `bcrypt` packages to project template.
+- add `djplusconfig` command to generate config file called `djplus.json`.
+- Implement `login`, `logout` and `signup` views in `auth` app.
+### Breaking Change
+- Change `User` model from `auth` app.
+- Change `Session` model from `auth` app.
+- Remove all migration files and create another one.
+
 ## [1.1.1] - 2022-06-29
 ### Fixed
 - Fixed the error of uninstalled packages with the help of `requirements` in the project template.
@@ -27,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2022-04-09
 Initial release
 
+[2.0.0]: https://github.com/githashem/djplus/compare/v1.1.1...v2.0.0
 [1.1.1]: https://github.com/githashem/djplus/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/githashem/djplus/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/githashem/djplus/compare/v0.1.0...v1.0.0